+++ internal/auth/jwt.go
@@ -5,8 +5,6 @@ import (
 	"encoding/hex"
 	"errors"
 	"log"
-	"net/http"
-	"strings"
 	"time"
 
 	"github.com/golang-jwt/jwt/v5"
@@ -61,18 +59,6 @@ func ValidateJWT(tokenString, tokenSecret string) (uuid.UUID, error) {
 		return uuid.Nil, err
 	}
 	return id, nil
-}
-
-func GetBearerToken(headers http.Header) (string, error) {
-	authHeader := headers.Get("Authorization")
-	if authHeader == "" {
-		return "", errors.New("no Authorization header found")
-	}
-	tokenString, ok := strings.CutPrefix(authHeader, "Bearer ")
-	if !ok {
-		return "", errors.New("no auth token found")
-	}
-	return tokenString, nil
 }
 
 func MakeRefreshToken() (string, error) {
